Put Location class into it's own file
