+++ iris/api/core.py
@@ -25,6 +25,7 @@ import pytesseract
 from errors import *
 from helpers.image_remove_noise import process_image_for_ocr, OCR_IMAGE_SIZE
 from helpers.parse_args import parse_args
+from location import Location
 
 try:
     import Image
@@ -683,7 +684,6 @@ def iris_image_match_template(needle, haystack, precision, threshold=None):
     """
     is_multiple = threshold is not None
 
-    from iris.api.core import Location
     try:
         res = cv2.matchTemplate(np.array(needle), np.array(haystack), FIND_METHOD)
     except:
@@ -770,59 +770,6 @@ class Pattern(object):
 
     def gray_image(self):
         return self.gray_image
-
-
-class Location(object):
-    def __init__(self, x=0, y=0):
-        self._x = x
-        self._y = y
-
-    def getX(self):
-        return self._x
-
-    def setX(self, new_x):
-        if isinstance(new_x, int):
-            self._x = new_x
-        else:
-            raise ValueError(INVALID_NUMERIC_INPUT)
-
-    x = property(getX, setX)
-
-    def getY(self):
-        return self._y
-
-    def setY(self, new_y):
-        if isinstance(new_y, int):
-            self._y = new_y
-        else:
-            raise ValueError(INVALID_NUMERIC_INPUT)
-
-    y = property(getY, setY)
-
-    def setLocation(self, x=0, y=0):
-        self._x = x
-        self._y = y
-
-    def offset(self, away_x, away_y):
-        new_x = int(self.x + away_x)
-        new_y = int(self.y + away_y)
-        return Location(new_x, new_y)
-
-    def above(self, away_y):
-        new_y = int(self.y - away_y)
-        return Location(self.x, new_y)
-
-    def below(self, away_y):
-        new_y = int(self.y + away_y)
-        return Location(new_y)
-
-    def left(self, away_x):
-        new_x = int(self.x - away_x)
-        return Location(new_x, self.y)
-
-    def right(self, away_x):
-        new_x = int(self.x + away_x)
-        return Location(new_x, self.y)
 
 
 class Region(object):
