+++ iris/api/core_helper.py
@@ -1,0 +1,2 @@
+INVALID_GENERIC_INPUT = 'Invalid input'
+INVALID_NUMERIC_INPUT = 'Expected numeric value'
