+++ iris/api/location.py
@@ -1,0 +1,54 @@
+from core_helper import INVALID_NUMERIC_INPUT
+
+
+class Location(object):
+    def __init__(self, x=0, y=0):
+        self._x = x
+        self._y = y
+
+    def getX(self):
+        return self._x
+
+    def setX(self, new_x):
+        if isinstance(new_x, int):
+            self._x = new_x
+        else:
+            raise ValueError(INVALID_NUMERIC_INPUT)
+
+    x = property(getX, setX)
+
+    def getY(self):
+        return self._y
+
+    def setY(self, new_y):
+        if isinstance(new_y, int):
+            self._y = new_y
+        else:
+            raise ValueError(INVALID_NUMERIC_INPUT)
+
+    y = property(getY, setY)
+
+    def setLocation(self, x=0, y=0):
+        self._x = x
+        self._y = y
+
+    def offset(self, away_x, away_y):
+        new_x = int(self.x + away_x)
+        new_y = int(self.y + away_y)
+        return Location(new_x, new_y)
+
+    def above(self, away_y):
+        new_y = int(self.y - away_y)
+        return Location(self.x, new_y)
+
+    def below(self, away_y):
+        new_y = int(self.y + away_y)
+        return Location(new_y)
+
+    def left(self, away_x):
+        new_x = int(self.x - away_x)
+        return Location(new_x, self.y)
+
+    def right(self, away_x):
+        new_x = int(self.x + away_x)
+        return Location(new_x, self.y)
